Refactor loop indices and type handling for improved safety and consistency
- Replaced `int` with `size_t` in loop indices across `CDlgGrep.cpp`, `CDlgOpenFile_CommonFileDialog.cpp`, `CDlgOpenFile_CommonItemDialog.cpp`, and `CPropTypesWindow.cpp`, enhancing compatibility with larger datasets and ensuring safer memory management.
- Updated documentation in AGENTS.md to clarify type usage distinctions, providing guidelines for new code and reinforcing overall type consistency across the codebase.

+++ sakura_core/dlg/CDlgGrep.cpp
@@ -632,10 +632,10 @@ void CDlgGrep::SetData( void )
 		HWND	hWndCombo = GetItemHwnd( IDC_COMBO_CHARSET );
 		int nCurIdx = ApiWrap::Combo_GetCurSel( hWndCombo );
 		CCodeTypesForCombobox cCodeTypes;
-		for (int nIdx = 0; nIdx < int(cCodeTypes.GetCount()); ++nIdx) {
-			const auto nCharSet = (ECodeType)ApiWrap::Combo_GetItemData( hWndCombo, nIdx );
+		for (size_t nIdx = 0; nIdx < cCodeTypes.GetCount(); ++nIdx) {
+			const auto nCharSet = (ECodeType)ApiWrap::Combo_GetItemData( hWndCombo, static_cast<int>(nIdx) );
 			if( nCharSet == m_nGrepCharSet ){
-				nCurIdx = nIdx;
+				nCurIdx = static_cast<int>(nIdx);
 			}
 		}
 		if( nCurIdx != -1 ){
